Issue is occuring using --checkpoint-open-files for ckpting open files
Consider a scenario in which file was opened for writing and however
its got deleted in between prior to taking dmtcp checkpoint. fds are
still opened for process but file is actually deleted. This is not
handled and dmtcp checkpoint exit as it tries to checkpoint the deleted
file. This was handled in previous within handleUnlinked file function
but code was cleaned.

Note: This has been committed in master already, so porting code to
2.5 branch too